Add some debugging around mysterious process launch failures.
The scenario that was causing this was the `$DEPS` being too long, due to using `$(classpath)`. This outputs the actual process params of the failing process.
Also fix `.close()`ing a `WorkerProcess` that hasn't been launched (which happens if the launch fails).

Separately, `assert`s in java are disabled by default (needs `-ea` in the cmdline), so perhaps there should be a lint rule
to use `Preconditions` instead?